Generate list of achievements and holes dynamically in /library
