+++ graphs/index.html
@@ -65,7 +65,9 @@
 </body>
 <script type="module">
   import "https://cdn.jsdelivr.net/npm/chart.js@2.9.4/dist/Chart.min.js";
-  import { sumAchievements, sumHoles, sumTalkNames, ACHIEVEMENTS, HOLES } from "../lib/stats.js";
+  import { ACHIEVEMENTS } from "../lib/achievements.js";
+  import { HOLES } from "../lib/holes.js";
+  import { sumAchievements, sumHoles, sumTalkNames } from "../lib/stats.js";
   import { useJSON } from "../lib/useJSON.js";
   import { useActiveLink } from "../lib/useActiveLink.js";
