+++ lib/stats.js
@@ -1,3 +1,6 @@
+import { HOLES } from "./holes.js"
+import { ACHIEVEMENTS } from "./achievements.js"
+
 export function sumAchievements(achievements) {
   return Object.values(achievements).reduce(
     (acc, achievement) => acc + achievement,
@@ -214,32 +217,6 @@ function parseArgs(keys, args) {
       })
   )
 }
-
-export const ACHIEVEMENTS = {
-  WaylandWithFallback: "--wayland-with-fallback",
-  NoFilesystemAll: "--no-filesystem-all",
-  NoFilesystemUnknown: "--no-filesystem-unknown",
-  NoFilesystem: "--no-filesystem",
-  NoDeviceAll: "--no-device-all",
-  NoDevice: "--no-device",
-  NoPulseaudio: "--no-pulseaudio",
-  NoNameTalk: "--no-name-talk",
-  NoNameOwn: "--no-name-own"
-};
-
-export const HOLES = {
-  X11: "--x11",
-  FilesystemAll: "--filesystem-all",
-  FilesystemUnknown: "--filesystem-unknown",
-  FilesystemWellKnown: "--filesystem-well-known",
-  DeviceAll: "--device-all",
-  DeviceWellKnown: "--device-well-known",
-  Pulseaudio: "--pulseaudio",
-  TalksSessionName: "--talks-session-name",
-  TalksSystemName: "--talks-system-name",
-  OwnsSessionName: "--owns-session-name",
-  OwnsSystemName: "--owns-system-name",
-};
 
 const WELL_KNOWN_XDG_RUN = "xdg-run";
